set up basic case study structure
