--- app/components/Flash.tsx
@@ -1,10 +1,0 @@
-export default function Flash({ className = '' }) {
-	return (
-		<svg className={className} fill="none" viewBox="0 0 433.32 469.5" xmlns="http://www.w3.org/2000/svg">
-			<path
-				className="fill-green mix-blend-multiply dark:mix-blend-lighten opacity-70"
-				d="M174.68,101.38c15.33-34.03,23.86-48.34,42.17-75.04,13.62,42.41,20.81,58.43,37.18,88.12,45.18-8.99,90.46-19.51,135.06-31.43-38.88,25.72-52.55,38.05-78.71,69.38,43.24,8.5,68.08,10.65,122.93,5.85-47.91,20.77-95.99,41.27-143.39,63.17,4.03,51.06.3,103.16-2.28,154.62-25.33-29.7-52.46-58.21-79.63-86.43-43.57,58.88-84.19,119.47-125,179.89,17.08-71.81,35.15-143.45,50.38-215.49-50.77-7.77-92.49-34.98-133.41-60.23,41.7-9.69,85-24.52,124.17-42.19C104.36,101.34,85.61,50.91,67.52,0c32.22,36.49,64.31,74.77,107.16,101.38Z"
-			/>
-		</svg>
-	)
-}
