--- app/components/GridLines.tsx
@@ -1,163 +1,0 @@
-export default function GridLines({ className = '' }) {
-	const lineClassName = 'fill-none stroke-indigo dark:stroke-teal stroke-[0.25]'
-	return (
-		<svg className={className} fill="none" viewBox="0 0 780 780" xmlns="http://www.w3.org/2000/svg">
-			<line className={lineClassName} x1="175.5" x2="175.5" y2="390" />
-			<line className={lineClassName} x1="156" x2="156" y2="390" />
-			<line className={lineClassName} x1="136.5" x2="136.5" y2="390" />
-			<line className={lineClassName} x1="117" x2="117" y2="390" />
-			<line className={lineClassName} x1="97.5" x2="97.5" y2="390" />
-			<line className={lineClassName} x1="78" x2="78" y2="390" />
-			<line className={lineClassName} x1="58.5" x2="58.5" y2="390" />
-			<line className={lineClassName} x1="39" x2="39" y2="390" />
-			<line className={lineClassName} x1="19.5" x2="19.5" y2="390" />
-			<line className={lineClassName} x1="351" x2="351" y2="390" />
-			<line className={lineClassName} x1="370.5" x2="370.5" y2="390" />
-			<line className={lineClassName} x1="331.5" x2="331.5" y2="390" />
-			<line className={lineClassName} x1="312" x2="312" y2="390" />
-			<line className={lineClassName} x1="292.5" x2="292.5" y2="390" />
-			<line className={lineClassName} x1="273" x2="273" y2="390" />
-			<line className={lineClassName} x1="253.5" x2="253.5" y2="390" />
-			<line className={lineClassName} x1="234" x2="234" y2="390" />
-			<line className={lineClassName} x1="214.5" x2="214.5" y2="390" />
-			<line className={lineClassName} x1="195" x2="195" y2="390" />
-			<line className={lineClassName} x1="390" y1="175.5" x2="0" y2="175.5" />
-			<line className={lineClassName} x1="390" y1="156" x2="0" y2="156" />
-			<line className={lineClassName} x1="390" y1="136.5" x2="0" y2="136.5" />
-			<line className={lineClassName} x1="390" y1="117" x2="0" y2="117" />
-			<line className={lineClassName} x1="390" y1="97.5" x2="0" y2="97.5" />
-			<line className={lineClassName} x1="390" y1="78" x2="0" y2="78" />
-			<line className={lineClassName} x1="390" y1="58.5" x2="0" y2="58.5" />
-			<line className={lineClassName} x1="390" y1="39" x2="0" y2="39" />
-			<line className={lineClassName} x1="390" y1="19.5" x2="0" y2="19.5" />
-			<line className={lineClassName} x1="390" y1="351" x2="0" y2="351" />
-			<line className={lineClassName} x1="390" y1="370.5" x2="0" y2="370.5" />
-			<line className={lineClassName} x1="390" y1="331.5" x2="0" y2="331.5" />
-			<line className={lineClassName} x1="390" y1="312" x2="0" y2="312" />
-			<line className={lineClassName} x1="390" y1="292.5" x2="0" y2="292.5" />
-			<line className={lineClassName} x1="390" y1="273" x2="0" y2="273" />
-			<line className={lineClassName} x1="390" y1="253.5" x2="0" y2="253.5" />
-			<line className={lineClassName} x1="390" y1="234" x2="0" y2="234" />
-			<line className={lineClassName} x1="390" y1="214.5" x2="0" y2="214.5" />
-			<line className={lineClassName} x1="390" y1="195" x2="0" y2="195" />
-			<line className={lineClassName} x1="565.5" x2="565.5" y2="390" />
-			<line className={lineClassName} x1="546" x2="546" y2="390" />
-			<line className={lineClassName} x1="526.5" x2="526.5" y2="390" />
-			<line className={lineClassName} x1="507" x2="507" y2="390" />
-			<line className={lineClassName} x1="487.5" x2="487.5" y2="390" />
-			<line className={lineClassName} x1="468" x2="468" y2="390" />
-			<line className={lineClassName} x1="448.5" x2="448.5" y2="390" />
-			<line className={lineClassName} x1="429" x2="429" y2="390" />
-			<line className={lineClassName} x1="409.5" x2="409.5" y2="390" />
-			<line className={lineClassName} x1="741" x2="741" y2="390" />
-			<line className={lineClassName} x1="760.5" x2="760.5" y2="390" />
-			<line className={lineClassName} x1="721.5" x2="721.5" y2="390" />
-			<line className={lineClassName} x1="702" x2="702" y2="390" />
-			<line className={lineClassName} x1="682.5" x2="682.5" y2="390" />
-			<line className={lineClassName} x1="663" x2="663" y2="390" />
-			<line className={lineClassName} x1="643.5" x2="643.5" y2="390" />
-			<line className={lineClassName} x1="624" x2="624" y2="390" />
-			<line className={lineClassName} x1="604.5" x2="604.5" y2="390" />
-			<line className={lineClassName} x1="585" x2="585" y2="390" />
-			<line className={lineClassName} x1="780" y1="175.5" x2="390" y2="175.5" />
-			<line className={lineClassName} x1="780" y1="156" x2="390" y2="156" />
-			<line className={lineClassName} x1="780" y1="136.5" x2="390" y2="136.5" />
-			<line className={lineClassName} x1="780" y1="117" x2="390" y2="117" />
-			<line className={lineClassName} x1="780" y1="97.5" x2="390" y2="97.5" />
-			<line className={lineClassName} x1="780" y1="78" x2="390" y2="78" />
-			<line className={lineClassName} x1="780" y1="58.5" x2="390" y2="58.5" />
-			<line className={lineClassName} x1="780" y1="39" x2="390" y2="39" />
-			<line className={lineClassName} x1="780" y1="19.5" x2="390" y2="19.5" />
-			<line className={lineClassName} x1="780" y1="351" x2="390" y2="351" />
-			<line className={lineClassName} x1="780" y1="370.5" x2="390" y2="370.5" />
-			<line className={lineClassName} x1="390" y1="390" x2="0" y2="390" />
-			<line className={lineClassName} x1="780" y1="390" x2="390" y2="390" />
-			<line className={lineClassName} x1="780" y1="331.5" x2="390" y2="331.5" />
-			<line className={lineClassName} x1="780" y1="312" x2="390" y2="312" />
-			<line className={lineClassName} x1="780" y1="292.5" x2="390" y2="292.5" />
-			<line className={lineClassName} x1="780" y1="273" x2="390" y2="273" />
-			<line className={lineClassName} x1="780" y1="253.5" x2="390" y2="253.5" />
-			<line className={lineClassName} x1="780" y1="234" x2="390" y2="234" />
-			<line className={lineClassName} x1="780" y1="214.5" x2="390" y2="214.5" />
-			<line className={lineClassName} x1="780" y1="195" x2="390" y2="195" />
-			<line className={lineClassName} x1="175.5" y1="390" x2="175.5" y2="780" />
-			<line className={lineClassName} x1="156" y1="390" x2="156" y2="780" />
-			<line className={lineClassName} x1="136.5" y1="390" x2="136.5" y2="780" />
-			<line className={lineClassName} x1="117" y1="390" x2="117" y2="780" />
-			<line className={lineClassName} x1="97.5" y1="390" x2="97.5" y2="780" />
-			<line className={lineClassName} x1="78" y1="390" x2="78" y2="780" />
-			<line className={lineClassName} x1="58.5" y1="390" x2="58.5" y2="780" />
-			<line className={lineClassName} x1="39" y1="390" x2="39" y2="780" />
-			<line className={lineClassName} x1="19.5" y1="390" x2="19.5" y2="780" />
-			<line className={lineClassName} x1="351" y1="390" x2="351" y2="780" />
-			<line className={lineClassName} x1="370.5" y1="390" x2="370.5" y2="780" />
-			<line className={lineClassName} x1="390" x2="390" y2="390" />
-			<line className={lineClassName} x1="390" y1="390" x2="390" y2="780" />
-			<line className={lineClassName} x1="331.5" y1="390" x2="331.5" y2="780" />
-			<line className={lineClassName} x1="312" y1="390" x2="312" y2="780" />
-			<line className={lineClassName} x1="292.5" y1="390" x2="292.5" y2="780" />
-			<line className={lineClassName} x1="273" y1="390" x2="273" y2="780" />
-			<line className={lineClassName} x1="253.5" y1="390" x2="253.5" y2="780" />
-			<line className={lineClassName} x1="234" y1="390" x2="234" y2="780" />
-			<line className={lineClassName} x1="214.5" y1="390" x2="214.5" y2="780" />
-			<line className={lineClassName} x1="195" y1="390" x2="195" y2="780" />
-			<line className={lineClassName} x1="390" y1="565.5" x2="0" y2="565.5" />
-			<line className={lineClassName} x1="390" y1="546" x2="0" y2="546" />
-			<line className={lineClassName} x1="390" y1="526.5" x2="0" y2="526.5" />
-			<line className={lineClassName} x1="390" y1="507" x2="0" y2="507" />
-			<line className={lineClassName} x1="390" y1="487.5" x2="0" y2="487.5" />
-			<line className={lineClassName} x1="390" y1="468" x2="0" y2="468" />
-			<line className={lineClassName} x1="390" y1="448.5" x2="0" y2="448.5" />
-			<line className={lineClassName} x1="390" y1="429" x2="0" y2="429" />
-			<line className={lineClassName} x1="390" y1="409.5" x2="0" y2="409.5" />
-			<line className={lineClassName} x1="390" y1="741" x2="0" y2="741" />
-			<line className={lineClassName} x1="390" y1="760.5" x2="0" y2="760.5" />
-			<line className={lineClassName} x1="390" y1="721.5" x2="0" y2="721.5" />
-			<line className={lineClassName} x1="390" y1="702" x2="0" y2="702" />
-			<line className={lineClassName} x1="390" y1="682.5" x2="0" y2="682.5" />
-			<line className={lineClassName} x1="390" y1="663" x2="0" y2="663" />
-			<line className={lineClassName} x1="390" y1="643.5" x2="0" y2="643.5" />
-			<line className={lineClassName} x1="390" y1="624" x2="0" y2="624" />
-			<line className={lineClassName} x1="390" y1="604.5" x2="0" y2="604.5" />
-			<line className={lineClassName} x1="390" y1="585" x2="0" y2="585" />
-			<line className={lineClassName} x1="565.5" y1="390" x2="565.5" y2="780" />
-			<line className={lineClassName} x1="546" y1="390" x2="546" y2="780" />
-			<line className={lineClassName} x1="526.5" y1="390" x2="526.5" y2="780" />
-			<line className={lineClassName} x1="507" y1="390" x2="507" y2="780" />
-			<line className={lineClassName} x1="487.5" y1="390" x2="487.5" y2="780" />
-			<line className={lineClassName} x1="468" y1="390" x2="468" y2="780" />
-			<line className={lineClassName} x1="448.5" y1="390" x2="448.5" y2="780" />
-			<line className={lineClassName} x1="429" y1="390" x2="429" y2="780" />
-			<line className={lineClassName} x1="409.5" y1="390" x2="409.5" y2="780" />
-			<line className={lineClassName} x1="741" y1="390" x2="741" y2="780" />
-			<line className={lineClassName} x1="760.5" y1="390" x2="760.5" y2="780" />
-			<line className={lineClassName} x1="721.5" y1="390" x2="721.5" y2="780" />
-			<line className={lineClassName} x1="702" y1="390" x2="702" y2="780" />
-			<line className={lineClassName} x1="682.5" y1="390" x2="682.5" y2="780" />
-			<line className={lineClassName} x1="663" y1="390" x2="663" y2="780" />
-			<line className={lineClassName} x1="643.5" y1="390" x2="643.5" y2="780" />
-			<line className={lineClassName} x1="624" y1="390" x2="624" y2="780" />
-			<line className={lineClassName} x1="604.5" y1="390" x2="604.5" y2="780" />
-			<line className={lineClassName} x1="585" y1="390" x2="585" y2="780" />
-			<line className={lineClassName} x1="780" y1="565.5" x2="390" y2="565.5" />
-			<line className={lineClassName} x1="780" y1="546" x2="390" y2="546" />
-			<line className={lineClassName} x1="780" y1="526.5" x2="390" y2="526.5" />
-			<line className={lineClassName} x1="780" y1="507" x2="390" y2="507" />
-			<line className={lineClassName} x1="780" y1="487.5" x2="390" y2="487.5" />
-			<line className={lineClassName} x1="780" y1="468" x2="390" y2="468" />
-			<line className={lineClassName} x1="780" y1="448.5" x2="390" y2="448.5" />
-			<line className={lineClassName} x1="780" y1="429" x2="390" y2="429" />
-			<line className={lineClassName} x1="780" y1="409.5" x2="390" y2="409.5" />
-			<line className={lineClassName} x1="780" y1="741" x2="390" y2="741" />
-			<line className={lineClassName} x1="780" y1="760.5" x2="390" y2="760.5" />
-			<line className={lineClassName} x1="780" y1="721.5" x2="390" y2="721.5" />
-			<line className={lineClassName} x1="780" y1="702" x2="390" y2="702" />
-			<line className={lineClassName} x1="780" y1="682.5" x2="390" y2="682.5" />
-			<line className={lineClassName} x1="780" y1="663" x2="390" y2="663" />
-			<line className={lineClassName} x1="780" y1="643.5" x2="390" y2="643.5" />
-			<line className={lineClassName} x1="780" y1="624" x2="390" y2="624" />
-			<line className={lineClassName} x1="780" y1="604.5" x2="390" y2="604.5" />
-			<line className={lineClassName} x1="780" y1="585" x2="390" y2="585" />
-		</svg>
-	)
-}
