--- app/components/HousePlantTall.tsx
@@ -1,12 +1,0 @@
-export default function Circle({ className = '' }) {
-	return (
-		<svg className={className} fill="none" viewBox="0 0 285.73 285.73" xmlns="http://www.w3.org/2000/svg">
-			<circle
-				className="fill-red mix-blend-hard-light opacity-70 dark:opacity-50"
-				cx="142.87"
-				cy="142.87"
-				r="142.87"
-			/>
-		</svg>
-	)
-}
