+++ app/components/LandingSection.tsx
@@ -1,9 +1,9 @@
 import Image from 'next/image'
-import GridLines from '@/components/GridLines'
-import Blob from '@/components/Blob'
-import Flash from '@/components/Flash'
-import LinedBlob from '@/components/LinedBlob'
-import Circle from '@/components/Circle'
+import GridLines from '@/lib/GridLines'
+import Blob from '@/lib/Blob'
+import Flash from '@/lib/Flash'
+import LinedBlob from '@/lib/LinedBlob'
+import Circle from '@/lib/Circle'
 import macbook from '@public/macbook.svg'
 import housePlantTall from '@public/house-plant-tall.svg'
 import housePlantShort from '@public/house-plant-short.svg'
