--- app/components/LinedBlob.tsx
@@ -1,67 +1,0 @@
-export default function LinedBlob({ className = '' }) {
-	const pathClassName = 'fill-yellow mix-blend-multiply dark:mix-blend-hard-light dark:opacity-20'
-	return (
-		<svg className={className} fill="none" viewBox="0 0 754.23 754.23" xmlns="http://www.w3.org/2000/svg">
-			<path
-				className={pathClassName}
-				d="M100.57,271.64c45.38,0,68.34-29.39,92.64-60.5,12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19s46.39,7.26,65.83,22.19c16.73,12.85,29.91,29.72,42.65,46.03,24.31,31.11,47.26,60.5,92.64,60.5s68.33-29.39,92.64-60.5c12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19,2.84,0,5.63.11,8.39.3-6.36-12.06-13.77-23.76-22.42-35.03-36.35,5.4-56.95,31.74-78.61,59.47-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.83-22.19c-16.73-12.85-29.91-29.72-42.65-46.03-24.31-31.11-47.26-60.5-92.64-60.5s-68.34,29.39-92.64,60.5c-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19-23.46,0-43.96-6.48-62.54-19.76-1.09,1.58-2.17,3.16-3.23,4.76-2.88,4.36-5.56,8.81-8.07,13.34,18.73,21.14,40.08,37.41,73.84,37.41Z"
-			/>
-			<path
-				className={pathClassName}
-				d="M367.54,53.37c16.73,12.85,29.91,29.72,42.65,46.03,24.31,31.11,47.26,60.5,92.64,60.5s68.33-29.39,92.64-60.5c12.03-15.4,24.46-31.3,39.88-43.84-5.09-3.64-10.29-7.1-15.59-10.36-2.82,3.52-5.62,7.1-8.44,10.71-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.83-22.19c-16.73-12.85-29.91-29.72-42.65-46.03-12.99-16.63-25.61-32.77-41.46-44.19-21.4,5.2-42.43,11.99-62.98,20.04,3.85-.38,7.78-.59,11.81-.59,24.85,0,46.39,7.26,65.83,22.19Z"
-			/>
-			<path
-				className={pathClassName}
-				d="M100.57,215.77c45.38,0,68.34-29.39,92.64-60.5,12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19s46.39,7.26,65.83,22.19c16.73,12.85,29.91,29.72,42.65,46.03,24.31,31.11,47.26,60.5,92.64,60.5s68.33-29.39,92.64-60.5c12.75-16.32,25.93-33.19,42.65-46.03,11.53-8.86,23.81-14.99,37.07-18.54-7.42-8.03-15.24-15.48-23.41-22.37-15.39,11.37-27.74,27.17-40.46,43.45-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.83-22.19c-16.73-12.85-29.91-29.72-42.65-46.03-24.31-31.11-47.26-60.5-92.64-60.5s-68.34,29.39-92.64,60.5c-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19-10.58,0-20.56-1.33-30.04-3.99-7.22,7.71-14.08,15.67-20.53,23.87,13.67,9.69,29.76,15.88,50.57,15.88Z"
-			/>
-			<path
-				className={pathClassName}
-				d="M100.57,159.9c45.38,0,68.34-29.39,92.64-60.5,10.34-13.23,20.96-26.82,33.52-38.34-4.87,2.61-9.71,5.27-14.5,7.99-9.29,5.28-18.65,10.82-27.99,16.61-5.54,5.85-11.43,11.36-17.84,16.29-9.67,7.43-19.86,12.95-30.72,16.64-16.86,12.55-33.13,26.02-48.26,40.42,4.17.58,8.53.89,13.14.89Z"
-			/>
-			<path
-				className={pathClassName}
-				d="M100.57,439.24c45.38,0,68.34-29.39,92.64-60.5,12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19s46.39,7.26,65.83,22.19c16.73,12.85,29.91,29.72,42.65,46.03,24.31,31.11,47.26,60.5,92.64,60.5s68.33-29.39,92.64-60.5c12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19,17.71,0,33.74,3.71,48.52,11.23-.72-11.45-1.57-22.87-2.64-34.24-12.65-7.88-27.41-12.75-45.89-12.75-45.38,0-68.33,29.39-92.64,60.5-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.83-22.19c-16.73-12.85-29.91-29.72-42.65-46.03-24.31-31.11-47.26-60.5-92.64-60.5s-68.34,29.39-92.64,60.5c-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.84-22.19c-13.18-10.12-24.16-22.75-34.46-35.63-.25,7.67-.37,15.42-.16,23.15,2.64,3.31,5.24,6.63,7.82,9.93,24.31,31.11,47.26,60.5,92.64,60.5Z"
-			/>
-			<path
-				className={pathClassName}
-				d="M443.69.02c-7.68.08-15.34.4-22.97.95,20.88,25.69,43.29,47.19,82.1,47.19,28.53,0,48.2-11.62,64.92-27.99-12.2-4.45-24.72-8.13-37.47-11.07-8.7,2.2-17.82,3.3-27.44,3.3-18.61,0-35.35-4.07-50.75-12.38-2.79-.03-5.59-.03-8.38,0Z"
-			/>
-			<path
-				className={pathClassName}
-				d="M436.99,46.08c-15.58-11.96-28.08-27.42-40.02-42.66-6.36.85-12.7,1.85-19.01,3,12,11.25,22.24,24.35,32.22,37.12,24.31,31.11,47.26,60.5,92.64,60.5s68.33-29.39,92.64-60.5c2.17-2.77,4.35-5.56,6.56-8.35-4.66-2.41-9.38-4.69-14.16-6.85-5.94,6.39-12.26,12.41-19.2,17.74-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.83-22.19Z"
-			/>
-			<path
-				className={pathClassName}
-				d="M611.32,447c-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.83-22.19c-16.73-12.85-29.91-29.72-42.65-46.03-24.31-31.11-47.26-60.5-92.64-60.5s-68.34,29.39-92.64,60.5c-12.75,16.32-25.93,33.19-42.65,46.03-2.45,1.88-4.94,3.64-7.46,5.28,9.05.55,18.05,1.02,26.86,1.58,2.47-3.1,4.93-6.25,7.41-9.41,12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19s46.39,7.26,65.83,22.19c16.73,12.85,29.91,29.72,42.65,46.03,24.31,31.11,47.26,60.5,92.64,60.5s68.33-29.39,92.64-60.5c12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19,16.28,0,31.12,3.13,44.89,9.48,1.45-10.24,2.59-20.54,3.44-30.9-13.19-8.8-28.65-14.34-48.33-14.34-45.38,0-68.33,29.39-92.64,60.5Z"
-			/>
-			<path
-				className={pathClassName}
-				d="M100.57,327.5c45.38,0,68.34-29.39,92.64-60.5,12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19s46.39,7.26,65.83,22.19c16.73,12.85,29.91,29.72,42.65,46.03,24.31,31.11,47.26,60.5,92.64,60.5s68.33-29.39,92.64-60.5c12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19,11.18,0,21.68,1.48,31.62,4.46-3.54-13.07-7.82-25.89-13.05-38.39-5.76-1.18-11.92-1.82-18.58-1.82-45.38,0-68.33,29.39-92.64,60.5-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.83-22.19c-16.73-12.85-29.91-29.72-42.65-46.03-24.31-31.11-47.26-60.5-92.64-60.5s-68.34,29.39-92.64,60.5c-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.84-22.19c-6.25-4.8-12-10.16-17.43-15.85-2.36,5.65-4.47,11.39-6.31,17.22,23.21,29.61,46.11,56.58,89.58,56.58Z"
-			/>
-			<path
-				className={pathClassName}
-				d="M7.92,322.88c24.31,31.11,47.26,60.5,92.64,60.5s68.34-29.39,92.64-60.5c12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19s46.39,7.26,65.83,22.19c16.73,12.85,29.91,29.72,42.65,46.03,24.31,31.11,47.26,60.5,92.64,60.5s68.33-29.39,92.64-60.5c12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19,15.61,0,29.91,2.87,43.21,8.7-1.58-12.27-3.53-24.43-5.99-36.43-10.65-5.05-22.8-8.02-37.23-8.02-45.38,0-68.33,29.39-92.64,60.5-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.83-22.19c-16.73-12.85-29.91-29.72-42.65-46.03-24.31-31.11-47.26-60.5-92.64-60.5s-68.34,29.39-92.64,60.5c-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.84-22.19c-11.26-8.65-20.92-19.13-29.92-30.03-1.25,6.56-2.2,13.19-2.86,19.88,2.01,2.54,4,5.08,5.97,7.61Z"
-			/>
-			<path
-				className={pathClassName}
-				d="M114.4,494.11c36.47-5.34,57.11-31.72,78.81-59.5,12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19s46.39,7.26,65.83,22.19c16.73,12.85,29.91,29.72,42.65,46.03,24.31,31.11,47.26,60.5,92.64,60.5s68.33-29.39,92.64-60.5c12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19,18.2,0,34.61,3.91,49.73,11.86.36-8.45.54-16.93.54-25.43-.12-2.25-.24-4.5-.36-6.76-13.53-9.43-29.44-15.43-49.92-15.43-45.38,0-68.33,29.39-92.64,60.5-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.83-22.19c-16.73-12.85-29.91-29.72-42.65-46.03-24.31-31.11-47.26-60.5-92.64-60.5s-68.34,29.39-92.64,60.5c-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.84-22.19c-11.35-8.72-21.07-19.29-30.13-30.29,2.1,8.83,5.07,17.39,9.22,25.53,20.98,41.13,59,55.57,100.58,61.71Z"
-			/>
-			<path
-				className={pathClassName}
-				d="M703.96,498.24c-45.38,0-68.33,29.39-92.64,60.5-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.83-22.19c-16.73-12.85-29.91-29.72-42.65-46.03-24.31-31.11-47.26-60.5-92.64-60.5-18.39,0-33.09,4.83-45.7,12.63,16.26,5.04,31.56,12.35,44.79,23.14.3,0,.6-.02.9-.02,24.85,0,46.39,7.26,65.83,22.19,16.73,12.85,29.91,29.72,42.65,46.03,24.31,31.11,47.26,60.5,92.64,60.5s68.33-29.39,92.64-60.5c12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19,6.3,0,12.38.47,18.27,1.41,4.13-10.67,7.83-21.46,11.11-32.37-8.7-3.06-18.38-4.8-29.38-4.8Z"
-			/>
-			<path
-				className={pathClassName}
-				d="M703.96,442.37c-45.38,0-68.33,29.39-92.64,60.5-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.83-22.19c-16.73-12.85-29.91-29.72-42.65-46.03-24.31-31.11-47.26-60.5-92.64-60.5s-67.84,28.76-91.86,59.5c2.76.3,5.48.62,8.17.97,4.15.55,8.28,1.19,12.4,1.92,1.78-1.52,3.59-3.01,5.46-4.44,19.45-14.93,40.98-22.19,65.84-22.19s46.39,7.26,65.83,22.19c16.73,12.85,29.91,29.72,42.65,46.03,24.31,31.11,47.26,60.5,92.64,60.5s68.33-29.39,92.64-60.5c12.75-16.32,25.93-33.19,42.65-46.03,19.45-14.93,40.98-22.19,65.84-22.19,12.36,0,23.89,1.8,34.76,5.44,2.59-10.24,4.84-20.57,6.76-30.98-11.66-6.37-25.12-10.22-41.51-10.22Z"
-			/>
-			<path
-				className={pathClassName}
-				d="M703.96,554.11c-45.38,0-68.33,29.39-92.64,60.5-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.83-22.19c-16.73-12.85-29.91-29.72-42.65-46.03-20.48-26.22-40.01-51.21-72.81-58.42,9.84,13.27,17.56,27.91,24.58,42.94,7.44,3.38,14.56,7.67,21.42,12.94,16.73,12.85,29.91,29.72,42.65,46.03,24.31,31.11,47.26,60.5,92.64,60.5s68.33-29.39,92.64-60.5c12.75-16.32,25.93-33.19,42.65-46.03,17.46-13.41,36.62-20.61,58.34-21.95.68-1.22,1.36-2.45,2.03-3.66,5.74-10.44,10.96-21.04,15.74-31.79-3.3-.35-6.71-.55-10.28-.55Z"
-			/>
-			<path
-				className={pathClassName}
-				d="M611.32,670.48c-12.75,16.32-25.93,33.19-42.65,46.03-19.45,14.93-40.98,22.19-65.84,22.19s-46.39-7.26-65.83-22.19c-16.73-12.85-29.91-29.72-42.65-46.03-10.62-13.6-20.99-26.86-33.08-37.49,2.28,5.06,4.6,10.09,7.02,15.03,32.52,66.51,92.75,129.57,173.59,97.69,18.71-7.38,36.21-17.8,52.48-30.32.37-.48.74-.95,1.12-1.43,12.75-16.32,25.93-33.19,42.65-46.03,3.82-2.93,7.72-5.56,11.7-7.9,12.35-15.38,23.57-31.58,33.63-47.82-32.39,7.41-51.79,32.22-72.14,58.26Z"
-			/>
-		</svg>
-	)
-}
