+++ app/lib/Blob.tsx
@@ -1,0 +1,10 @@
+export default function Blob({ className = '' }) {
+	return (
+		<svg className={className} fill="none" viewBox="0 0 492.4 512.13" xmlns="http://www.w3.org/2000/svg">
+			<path
+				className="fill-purple mix-blend-multiply dark:mix-blend-hard-light opacity-70 dark:opacity-60"
+				d="M212.54,351.94c-2.59-1.76-4.72-3.22-6.71-4.57-61.86-3.8-75.52,109.04-141.5,83.84-46.86-17.9-93.71-116.42-41.14-151.14,23.07-15.23,59.58-29.98,65.98-60.59,6.38-30.46-27.6-51.85-33.94-79.31-2.95-12.78,2.66-30.67,13.86-38.01,11.9-7.8,29.83-4.21,42.8-2.1,29.32,4.77,64.07,14.89,89.96-5.27,25.16-19.59,29.54-53.77,50.09-76.91,19.93-22.44,52.76-25.09,72.09-.22,22.29,28.66,4.32,57.32-9.89,84.8-7.81,15.1-15.74,29.98-20.97,46.21-4.62,14.34-10.38,30.64-1.06,44.49,17.83,26.48,63.22,15.99,90.15,19.54,31.39,4.14,63.58,15.01,85.79,38.66,47.35,50.41,17.03,118.43-18.76,165.77-39.62,52.4-111.42,109.56-181.85,91.66-28.08-7.14-51.39-28.31-54.87-57.99-3.58-30.51,21.65-71.52-.06-98.84Z"
+			/>
+		</svg>
+	)
+}
