+++ app/utils/constants.ts
@@ -17,53 +17,4 @@ export const NAV_ROUTES = [
 	},
 ]
 
-export const PROJECTS_URL = '/projects'
-
-export const PROJECT_ROUTES = [
-	{
-		href: `${PROJECTS_URL}/tpc-rebrand`,
-		title: 'TPC: E-commerce Rebrand',
-	},
-	{
-		href: `${PROJECTS_URL}/tpc-component-library`,
-		title: 'TPC: Component Library',
-	},
-	{
-		href: `${PROJECTS_URL}/starbucks-s4-apps`,
-		title: 'Starbucks: Web Apps',
-	},
-	{
-		href: `${PROJECTS_URL}/wa-doh-vaccine-locator`,
-		title: 'WA DOH: Vaccine Locator',
-	},
-	{
-		href: `${PROJECTS_URL}/astarte-inventory-management`,
-		title: 'Astarte: Inventory Management System',
-	},
-	{
-		href: `${PROJECTS_URL}/paragon`,
-		title: 'Paragon: Luxury Real Estate Website',
-	},
-	{
-		href: `${PROJECTS_URL}/paragon-property-management`,
-		title: 'Paragon: Property Management System',
-	},
-	{
-		href: `${PROJECTS_URL}/mighty-vet-ux`,
-		title: 'Mighty Vet: Website Concept',
-	},
-	{
-		href: `${PROJECTS_URL}/wa-state-parks-camping-reservations`,
-		title: 'WA State Parks: Camping Reservation System',
-	},
-	{
-		href: `${PROJECTS_URL}/high-seas-portfolio`,
-		title: 'High Seas: Portfolio Website',
-	},
-	{
-		href: `${PROJECTS_URL}/ian-j-smith-portfolio`,
-		title: "ianjsmith.com: How It's Made",
-	},
-]
-
 export const WEB_FORM_KEY = '5602ac20-bcd3-48da-bd9e-51ddde3b01df'
